+++ CHANGELOG.md
@@ -33,6 +33,7 @@ versioned in accordance with [Semantic Versioning](https://semver.org/spec/v2.0.
 - Removed the 'Connector` suffix from the default connector name attribute and introduced `full_name`.
 - The `__config_key__` attribute has been renamed to `__default_name__`.
 - The identifier for connectors has been renamed to `name` for simplicitly and clarity.
+- The `ServoAssembly` class has been renamed to `Assembly`.
 
 ### Removed
 - The `durationpy` package has been removed in favor of a local implementation.